Add ability to play music and sound effects
Sounds and music are in a custom format I call NSF for Nekro sound
format, consisting of opcodes for audio register writes.

One sound effect per audio channel can be played at a time. For the
duration of the sfx, music will be silenced, but the register state will
be stored in reserve registers that will be copied back into the actual
registers when the sfx is done to resume the music.