Rearrange tilemap related code
Big cleanup that fixed a bug in tileset attribute indexing as well.
Level now has a tilemap as a separate thing, and the tilemap is dealt
with separately from other level stuff in tilemap related functions.
Actor stuff is currently commented out until I figure out what to do
with it (There are lots of plans)